Consolidate saved decks code to HomePagePlayMode
